+++ src/de/omegazirkel/risingworld/DiscordConnect.java
@@ -49,8 +49,11 @@ import de.omegazirkel.risingworld.tools.FileChangeListener;
 import de.omegazirkel.risingworld.tools.I18n;
 import de.omegazirkel.risingworld.tools.OZLogger;
 import de.omegazirkel.risingworld.tools.settings.PlayerPluginAdminSettings;
+import de.omegazirkel.risingworld.tools.ui.AssetManager;
+import de.omegazirkel.risingworld.tools.ui.MenuItem;
 import de.omegazirkel.risingworld.tools.ui.PlayerPluginSettingsOverlay;
 import de.omegazirkel.risingworld.tools.ui.PluginInfoStatusProviders;
+import de.omegazirkel.risingworld.tools.ui.PluginMenuManager;
 import net.risingworld.api.Plugin;
 import net.risingworld.api.Server;
 import net.risingworld.api.events.EventMethod;
@@ -130,14 +133,20 @@ public class DiscordConnect extends Plugin implements Listener, FileChangeListen
 		this.initialize();
 
 		// register plugin settings
+		AssetManager.loadIconFromPlugin(this, "icon-ki-discord-connect");
 		PlayerPluginSettingsOverlay.registerPlayerPluginSettings(new DiscordConnectPlayerPluginSettings(getDescription("version")));
 		PlayerPluginSettingsOverlay.registerPlayerPluginData(new DiscordConnectPlayerPluginData(getDescription("version")));
 		PlayerPluginSettingsOverlay.registerPlayerPluginAdminSettings(
 				new PlayerPluginAdminSettings(name, getDescription("version"), () -> s.adminSettingsEntries(),
 						s::initSettings));
-		PluginInfoStatusProviders
-				.registerProvider(new DiscordConnectPluginInfoStatusProvider(this, getDescription("version")));
-		logger().info("✅ " + this.getName() + " Plugin is enabled version:" + this.getDescription("version"));
+			PluginInfoStatusProviders
+					.registerProvider(new DiscordConnectPluginInfoStatusProvider(this, getDescription("version")));
+			PluginMenuManager.registerPluginMenu(new MenuItem(AssetManager.getIcon("icon-ki-discord-connect"),
+					"Discord Connect", player -> {
+						player.hideRadialMenu(true);
+						PluginInfoStatusProviders.show(player, name);
+					}));
+			logger().info("✅ " + this.getName() + " Plugin is enabled version:" + this.getDescription("version"));
 
 	}